+++ astro/forms/src/utils/constants.ts
@@ -1,0 +1,12 @@
+export enum FormFieldsWixIds {
+  teamName = "team_name",
+  ageGroup = "age_group_1",
+  skillLevel = "skill_level",
+  firstName = "first_name_1536",
+  lastName = "last_name_7eec",
+  email = "email_8874",
+  phone = "phone_f28a",
+  specialRequirements = "special_requirements_or_requests",
+}
+
+export const WIX_FORM_ID = "2fc1513a-7ad0-4a5e-a731-1d0fb6e0e7e1";
